# Avoid allocating on each call to rows.Columns