### Description

Calling rows.Columns() always allocates a new slice, even if it was already called once. Cache the slice on the first call and return the cached value on subsequent calls.

Also if we have a prepared statement, cache the column names in the statement the first time it gets executed and set them the next time the prepared statement gets executed (as is already done with the column information).

This also helps if rows.Columns() is never called directly by the user, as rows.Next() (sql.Rows, not driver.Rows) curently calls rows.Columns() internally on the first call.

I also looked at the lib/pq PostgreSQL driver for Go and they [seem to do the same thing](https://github.com/lib/pq/blob/165a3529e799da61ab10faed1fabff3662d6193f/conn.go#L1413-L1415).

In a small benchmark that selects 1 row with 1 string column, I get these numbers:
(DatabaseSQL uses a non-prepared query)

```
name                   old time/op    new time/op    delta
DatabaseSQL-8            20.4µs ± 4%    20.6µs ± 4%     ~             (p=0.548 n=5+5)
DatabaseSQLPrepared-8    20.8µs ± 7%    21.5µs ±14%     ~             (p=0.841 n=5+5)

name                   old alloc/op   new alloc/op   delta
DatabaseSQL-8              320B ± 0%      352B ± 0%  +10.00%          (p=0.008 n=5+5)
DatabaseSQLPrepared-8      304B ± 0%      320B ± 0%   +5.26%          (p=0.008 n=5+5)

name                   old allocs/op  new allocs/op  delta
DatabaseSQL-8              11.0 ± 0%      11.0 ± 0%     ~     (all samples are equal)
DatabaseSQLPrepared-8      10.0 ± 0%       9.0 ± 0%  -10.00%          (p=0.008 n=5+5)
```
### Checklist
- [x] Code compiles correctly
- [x] Created tests which fail without the change (if possible)
- [x] All tests passing
- [x] Extended the README / documentation, if necessary
- [x] Added myself / the copyright holder to the AUTHORS file