@methane 2016-08-25
LGTM

> Also if we have a prepared statement, cache the column names in the statement the first time it gets executed and set them the next time the prepared statement gets executed (as is already done with the column information).

I'm afraid that adding / removing column from another session may break the cache.
But column info caching is already there.  This PR doesn't make it worse.