@julienschmidt 2016-12-01
Or at least team members can reopen it then