@nuss-justin 2016-12-01
That worked :-) I also updated the benchmark in the PR for the Go 1.8 beta.