@methane 2017-05-12
Please merge master.

> Also if we have a prepared statement, cache the column names in the statement the first time it gets executed and set them the next time the prepared statement gets executed (as is already done with the column information).

The information cache is removed because prepared statement can return different columns.